add flags

+++ src/services/UserService.ts
@@ -3,6 +3,7 @@ import HttpStatusCodes from '@src/common/HttpStatusCodes';
 
 import UserRepo from '@src/repos/UserRepo';
 import { IUser } from '@src/models/User';
+import featureService from '@src/flags';
 
 
 // **** Variables **** //
@@ -23,6 +24,13 @@ function getAll(): Promise<IUser[]> {
  * Add one user.
  */
 function addOne(user: IUser): Promise<void> {
+  const allowSignups = featureService.isEnabled('allowSignups');
+  if (!allowSignups) {
+    throw new RouteError(
+      HttpStatusCodes.FORBIDDEN,
+      'Signups are not allowed',
+    );
+  }
   return UserRepo.add(user);
 }
 
@@ -52,8 +60,14 @@ async function _delete(id: number): Promise<void> {
       USER_NOT_FOUND_ERR,
     );
   }
-  // Delete user
-  return UserRepo.delete(id);
+
+  const hardDelete = featureService.isEnabled('hardDeleteUsers');
+  if (hardDelete) {
+    return UserRepo.delete(id);
+  }
+
+  // Soft delete user
+  return UserRepo.softDelete(id);
 }
